В main и в SalasManager все инты заменены на лонги

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
 
-        SalesManager SalesManager = new SalesManager(new int[]{2, 3, 3, 14, 10});
+        SalesManager SalesManager = new SalesManager(new long[]{2, 3, 3, 14, 10});
 
         System.out.println(SalesManager.max());
     }
